Remove "bonfire" and "pyre" code
"bonfire" was an embryonic implementation of the B2 service, and "pyre" was a
gRPC-based implementation of B2, with a RESTful gateway on top of it. This
commit removes both bonfire and pyre since neither of them is used, as far as
we know, and they pull in old dependencies that trigger spurious security
alerts.